Separate components for body and footer.
